Changes for blueos extension

+++ backend_py/src/services/cameras/ehd.py
@@ -31,7 +31,7 @@ class EHDDevice(Device):
         # Standard integer options
         options['bitrate'] = Option(
             self.cameras[2], '>I', xu.Unit.USR_ID, xu.Selector.USR_H264_CTRL, xu.Command.H264_BITRATE_CTRL, 'Bitrate', 
-            lambda bitrate: bitrate * 1000000, # convert to bps from mpbs 
+            lambda bitrate: int(bitrate * 1000000), # convert to bps from mpbs 
             lambda bitrate: bitrate / 1000000  # convert to mpbs from bps
         )
